Play 5 rounds of rock paper scissors
Prompts user to enter either rock, paper, or scissors up to 5 times.
Each game played will display wether you win or lose, not only that
but it will also display the current score in the game. Once the game
ends it'll display the final score for both user and computer